+++ src/terrama2/services/analysis/core/OperatorCache.hpp
@@ -51,10 +51,10 @@ namespace terrama2
         */
         struct OperatorCache
         {
-          int64_t index = -1; //!< Geometry index of the monitored object.
+          int32_t index = -1; //!< Geometry index of the monitored object.
           AnalysisHashCode analysisHashCode = 0; //!< Hashcode of current analysis.
-          int64_t row = -1; //!< Output raster row.
-          int64_t column = -1; //!< Output raster column.
+          int32_t row = -1; //!< Output raster row.
+          int32_t column = -1; //!< Output raster column.
           double sum = 0; //!< Result of the sum.
           double max = std::numeric_limits<double>::min(); //!< Maximum value.
           double min = std::numeric_limits<double>::max(); //!< Minimum value.
@@ -62,7 +62,7 @@ namespace terrama2
           double mean = 0; //!< Mean value.
           double standardDeviation = 0; //!< Standard deviation value.
           double variance = 0; //!< Standard deviation value.
-          uint64_t count = 0; //!< Count value.
+          uint32_t count = 0; //!< Count value.
         };
       } // end namespace core
     }   // end namespace analysis
